+++ geo/CHANGES.md
@@ -85,6 +85,8 @@
   - <https://github.com/georust/geo/pull/1317>
 - Add: PreparedGeometry::geometry and into_geometry to get at the inner geometry type.
   - <https://github.com/georust/geo/pull/1318>
+- Add: Clone and Debug implementations for PreparedGeometry.
+  - <https://github.com/georust/geo/pull/1324>
 
 ## 0.29.3 - 2024.12.03